fix: split test code
make it works on golang version lower than 1.18
